requirements.txt
pypdf
transformers
sentence_transformers
faiss_cpu
gradio
langchain
openai
huggingface_hub
tiktoken
python-dotenv